Analysis: Physics: Move manager_t to separate header in namespace ant::analysis::physics
