+++ src/analysis/CMakeLists.txt
@@ -27,6 +27,7 @@ set(ANALYSIS_SLOWCONTROL
 set(ANALYSIS_PHYSICS
   physics/Physics.cc
   physics/PhysicsManager.cc
+  physics/manager_t.h
   physics/test/TestPhysics.cc
   physics/pi0/DeltaPlusPhysics.cc
   physics/pi0/DeltaTreeGenerator.cc
