+++ src/analysis/physics/common/EventFilter.cc
@@ -31,7 +31,7 @@ EventFilter::EventFilter(const string& name, OptionsPtr opts):
 EventFilter::~EventFilter()
 {}
 
-void EventFilter::ProcessEvent(const TEvent& event, Physics::manager_t& manager)
+void EventFilter::ProcessEvent(const TEvent& event, manager_t& manager)
 {
     const auto& data = *(event.Reconstructed);
